Ticket #5669 - removed useless testcase, fixed som warnings
git-svn-id: file:///data/svn/enum/server/trunk@17844 b60c293d-5e0d-0410-87b2-fb9704ccd4b5

+++ src/fredlib/requests/request_manager.cc
@@ -856,7 +856,7 @@ ManagerImpl::i_getRequestCountUsers(
 
     std::auto_ptr<RequestCountInfo> info (new RequestCountInfo());
 
-    for(int i=0;i<res.size();++i) {
+    for(size_t i=0;i<res.size();++i) {
         std::string user_handle = (std::string)res[i][0];
         unsigned long long count = (unsigned long long)res[i][1];
